moved getting phases to App.tsx for use with Search Form everywhere
