+++ src/App.test.tsx
@@ -1,3 +1,4 @@
+import axios from 'axios';
 import mockI18next from '@/mocks/mockI18next';
 import { act, waitFor } from '@testing-library/react';
 import { renderWithProviders } from './utils/testUtils';
@@ -5,10 +6,25 @@ import App from './App';
 import { mockGetResponseProvider } from './utils/mockGetResponseProvider';
 import mockProjectClasses from './mocks/mockClasses';
 import { mockLocations } from './mocks/mockLocations';
-import { mockProjectCategories, mockResponsiblePersons } from './mocks/mockLists';
+import {
+  mockProjectCategories,
+  mockProjectPhases,
+  mockResponsiblePersons,
+} from './mocks/mockLists';
+import { mockError } from './mocks/mockError';
+import {
+  getProjectCategoriesThunk,
+  getProjectPhasesThunk,
+  getResponsiblePersonsThunk,
+} from './reducers/listsSlice';
+import { IError } from './interfaces/common';
+import { getClassesThunk } from './reducers/classSlice';
+import { getLocationsThunk } from './reducers/locationSlice';
 
 jest.mock('axios');
 jest.mock('react-i18next', () => mockI18next());
+
+const mockedAxios = axios as jest.Mocked<typeof axios>;
 
 describe('App', () => {
   beforeEach(async () => {
@@ -28,6 +44,7 @@ describe('App', () => {
 
     expect(lists.categories).toStrictEqual(mockProjectCategories.data);
     expect(lists.responsiblePersons).toStrictEqual(mockResponsiblePersons.data);
+    expect(lists.phase).toStrictEqual(mockProjectPhases.data);
     expect(classes.allClasses).toStrictEqual(mockProjectClasses.data);
     expect(classes.masterClasses.length).toBeGreaterThan(0);
     expect(classes.classes.length).toBeGreaterThan(0);
@@ -72,5 +89,55 @@ describe('App', () => {
       expect(getByText('error.pageNotFound')).toBeInTheDocument();
       expect(getByTestId('return-to-previous-btn')).toBeInTheDocument();
     });
+  });
+
+  it('catches a failed phase list fetch', async () => {
+    const { store } = renderResult;
+    mockedAxios.get.mockRejectedValueOnce(mockError);
+    await store.dispatch(getProjectPhasesThunk());
+
+    const storeError = store.getState().lists.error as IError;
+    expect(storeError.message).toBe(mockError.message);
+    expect(storeError.status).toBe(mockError.status);
+  });
+
+  it('catches a failed projectCategories list fetch', async () => {
+    const { store } = renderResult;
+    mockedAxios.get.mockRejectedValueOnce(mockError);
+    await store.dispatch(getProjectCategoriesThunk());
+
+    const storeError = store.getState().lists.error as IError;
+    expect(storeError.message).toBe(mockError.message);
+    expect(storeError.status).toBe(mockError.status);
+  });
+
+  it('catches a failed responsible persons fetch', async () => {
+    const { store } = renderResult;
+    mockedAxios.get.mockRejectedValueOnce(mockError);
+    await store.dispatch(getResponsiblePersonsThunk());
+
+    const storeError = store.getState().lists.error as IError;
+    expect(storeError.message).toBe(mockError.message);
+    expect(storeError.status).toBe(mockError.status);
+  });
+
+  it('catches a failed classes fetch', async () => {
+    const { store } = renderResult;
+    mockedAxios.get.mockRejectedValueOnce(mockError);
+    await store.dispatch(getClassesThunk());
+
+    const storeError = store.getState().class.error as IError;
+    expect(storeError.message).toBe(mockError.message);
+    expect(storeError.status).toBe(mockError.status);
+  });
+
+  it('catches a failed locations fetch', async () => {
+    const { store } = renderResult;
+    mockedAxios.get.mockRejectedValueOnce(mockError);
+    await store.dispatch(getLocationsThunk());
+
+    const storeError = store.getState().location.error as IError;
+    expect(storeError.message).toBe(mockError.message);
+    expect(storeError.status).toBe(mockError.status);
   });
 });
